Pointer Events mixins and Layers
PointerEvents enable developers to interact with objects in the world.
Layers are used to segment functional systems in to a number of different uses. At the moment, you can't mix these, but this may change (you may want to walk on an editable object for example).
LIGHT_LAYER
EVENT_LAYER
WALK_LAYER
COLLIDE_LAYER
AVATAR_LAYER