sayDeck/listenDeck
This may change, but it should look very much like this. This enables any card and surface to communicate via the parent. Thus, if a surface calls:
sayDeck('setColor', color) and a child DCard subclass is listening via:
listenDeck('setColor', this.setColor)
then the child DCard's color is changed! Very, very convenient.