Added static: flag to cards
When a card is set to static, it means the 3D model won't be moving around internally, so the object tree can be flattened. This can turn a complex object that may have hundreds of parts into just 2 or 3 which dramatically improves rendering performance and object management.

Also includes a bit of cleanup.